ENH: Add VTK 6 support 2/2
Issue #3514

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=23122
git-svn-id: http://svn.slicer.org/Slicer4/trunk@23122 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/vtkITK/vtkITKBSplineTransform.cxx
@@ -335,13 +335,13 @@ vtkITKBSplineTransform
     vtkErrorMacro( "need to call SetSplineOrder before SetParameters" );
     return;
     }
-  unsigned int N=Helper->GetNumberOfParameters();
+  size_t N=Helper->GetNumberOfParameters();
   if (N==0)
     {
     return;
     }
   double *paramDouble=new double[N];
-  for (unsigned int i=0; i<N; i++)
+  for (size_t i=0; i<N; i++)
   {
     paramDouble[i]=param[i];
   }
@@ -402,7 +402,7 @@ vtkITKBSplineTransform
     return;
     }
   double *paramDouble=new double[N];
-  for (unsigned int i=0; i<N; i++)
+  for (size_t i=0; i<N; i++)
   {
     paramDouble[i]=param[i];
   }
